refactor :ファイルの構造を見直し
