+++ js/main.js
@@ -1,9 +1,9 @@
 const submitButton = document.querySelector('.js_addTodo_submit');
 const todoTaskList = document.querySelector('.js_todoTask_list');
 const inputField = document.querySelector('.js_addTodo_input');
-import { Todo } from './todo.js';
-import toggleSubmitActive from './toggleSubmitActive.js';
-import dropdown from './dropdown.js';
+import { Todo } from './modules/todo.js';
+import toggleSubmitActive from './modules/toggleSubmitActive.js';
+import dropdown from './modules/dropdown.js';
 
 toggleSubmitActive();
 inputField.addEventListener('input', () => {
