+++ js/modules/textUtility.js
